feat!: add responsive label texts to CountryCodeSelector and easily internationalizable ITU-T E.164 compliant country data system (#15)
* feat!: add responsive default values for CountryCodeSelector

The CountryCodeSelector label is abbreviated and ISO code is used as a label for the selected country when the screen is smaller than the smallest breakpoint of the used theme. (NOTICE! Overridden by a later commit in this PR; shorter labels are used when the element width is smaller than a customizable threshold value.)

* fix: make demo-page install mui-country-code-selector from the repository

* feat!: new internationalizable country code data system with updated data

List of countries is now fetched using getCountries() function. List now contains all ITU-T E.164 assigned calling codes and the countries and subdivision areas are identified using ISO 3166 country codes. NOTICE! Due to the way Vite handles JSON keys during import, the dash has been left out of the ISO 3166-2 codes of the subdivision areas where the codes are used as keys.

* refactor: handling of component default values

* build(deps): remove css-mediaquery and audit fix vulnerabilities

* feat!: make CountryCodeSelector switch label length automatically when its size changes

* feat: edit the names of North and South Korea

* fix: AutocompleteProps typing that caused errors when using disableClearable

* docs: update README.md and package.json keywords

* feat!: update the content of the demo page to reflect the new features